- EPROTO may be defined in newer versions of FreeBSD (like 6.1) - fixed several compiling warnings on 64bit arch.
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@891 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/postgres/db_val.c
@@ -243,7 +243,7 @@ int str2valp(db_type_t _t, db_val_t* _v, char* _s, int _l, void *_p)
 
 	case DB_BLOB:
 		VAL_BLOB(_v).s = (char*)PQunescapeBytea((unsigned char*)_s,
-				(unsigned int*)&(VAL_BLOB(_v).len) );
+				(size_t*)&(VAL_BLOB(_v).len) );
 		VAL_TYPE(_v) = DB_BLOB;
 		DBG("DEBUG:postgres:str2valp: got blob len %d\n", _l);
 		return 0;
@@ -259,7 +259,7 @@ int val2str(db_val_t* _v, char* _s, int* _len)
 {
 	int l;
 	char *tmp_s;
-	int tmp_len;
+	size_t tmp_len;
 
 #ifdef PARANOID
 	if ((!_v) || (!_s) || (!_len) || (!*_len)) {
@@ -348,7 +348,7 @@ int val2str(db_val_t* _v, char* _s, int* _len)
 		} else {
 			*_s++ = '\'';
 			tmp_s = (char*)PQescapeBytea((unsigned char*)VAL_STRING(_v),
-					l, (unsigned int*)&tmp_len);
+					(size_t)l, (size_t*)&tmp_len);
 			memcpy(_s, tmp_s, tmp_len);
 			PQfreemem(tmp_s);
 			tmp_len = strlen(_s);
